requirements.txt
# Core
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Technical Analysis
ta>=0.11.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
tensorflow>=2.15.0

# NLP / LLM
openai>=1.10.0
google-generativeai>=0.3.0

# Database
sqlalchemy>=2.0.0

# HTTP
requests>=2.31.0
